# Add `set_trial_user_attrs` method for bulk upsert