--- a/optuna/trial/_frozen.py
+++ b/optuna/trial/_frozen.py
@@ -297,6 +297,9 @@ class FrozenTrial(BaseTrial):
     def set_user_attr(self, key: str, value: Any) -> None:
         self._user_attrs[key] = value
 
+    def set_user_attrs(self, attrs: dict[str, Any]) -> None:
+        self._user_attrs.update(attrs)
+
     @deprecated_func("3.1.0", "5.0.0")
     def set_system_attr(self, key: str, value: Any) -> None:
         self._system_attrs[key] = value